Moved to doing all IO at the beginning
This should hopefully increase performance, as it doesn't require
partial reading of the file - as all is done up front, and any further
access to the file is in grabbing the line from the vector stored inside
the File object.

Some other small changes - e.g. changing around how comments are parsed
in the Tokenizer - are also included.