CHanged to not header only
I've moved away from using header only files. The main reason
being that this should decouple compile time for specific items
without having to recompile everything that depends on it.

This has also been for optimisation efforts: It's allowed me
to look over all of the files as a whole to find out what can
be refactored and optimised, or modularised, and hopefully will
lead to some improvements in future.